npm install gulp-sass error
Many of us have tried running `npm install` and every time we do the gulp-sass fails to build and we get a serve error when we run `gulp serve` so running 
```
npm uninstall --save-dev gulp-sass
npm install --save-dev gulp-sass@2
```
Fixed the problems. Instead of people needing to search for the solution thought we would fix the bug by changing the gulp-sass version.